server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.0.0",
    "alembic[tz]>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.124.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "greenlet>=3.3.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
    "testcontainers>=4.13.3",
    "ty>=0.0.13",
    "types-pyjwt>=1.7.1",
]
